+++ changelog.md
@@ -1,6 +1,8 @@
 CURRENT:
 - Potental Fix for Collision Loop Sound Effects Persisting after Impact
 - Tweak Orbit View Muffling
+- Removed "MuffleChatterer" from LOWPASSFILTER in Settings.cfg
+- Added "AffectChatterer" under RSE_SETTINGS in Settings.cfg
 
 0.7.1
 - Fixed Music being Muffled